requirements.txt
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alpaca-trade-api==3.2.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
black==25.1.0
blinker==1.9.0
bs4==0.0.2
cachetools==6.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.2
cycler==0.12.1
deprecation==2.1.0
distro==1.9.0
etelemetry==0.3.1
feedparser==6.0.11
filelock==3.18.0
fitz==0.0.1.dev2
Flask==3.1.1
fonttools==4.58.4
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
looseversion==1.3.0
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
msgpack==1.0.3
multidict==6.5.0
mypy==1.15.0
mypy_extensions==1.1.0
narwhals==1.43.1
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pathlib==1.0.1
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
pluggy==1.6.0
propcache==0.3.2
protobuf==6.31.1
prov==2.0.2
puremagic==1.29
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pydot==4.0.1
PyMuPDF==1.26.1
pyparsing==3.2.3
pyright==1.1.401
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
trimesh==4.5.2
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
faiss-cpu==1.9.0
schedule==1.2.2
scikit-learn==1.5.2
sentence-transformers==3.3.1
psutil==6.1.1
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.1
rdflib==7.1.4
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
ruff==0.11.10
scipy==1.15.3
sgmllib3k==1.0.0
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
streamlit==1.46.0
sympy==1.14.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traits==7.0.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
watchdog==6.0.0
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.3
yarl==1.20.1

# Additional Dependencies for JARVIS v3.0 Autonomy
fastapi==0.104.1
uvicorn[standard]==0.24.0
croniter==2.0.1
cron-descriptor==1.4.0
joblib==1.3.2
dash==2.14.2
dash-bootstrap-components==1.5.0
textblob==0.17.1
vaderSentiment==3.3.2
newspaper3k==0.2.8
rich==13.7.0
colorama==0.4.6
orjson==3.9.10
loguru==0.7.2
python-multipart==0.0.6
